src/requirements.txt
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.10.15
chardet==3.0.4
dnspython==1.16.0
idna==2.7
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.20.0
robobrowser==0.5.3
six==1.11.0
urllib3==1.24.2
Werkzeug==0.15.3
